requirements.txt
django>=1.8
djangorestframework>=3.0
djangorestframework-filters
django-appconf==1.0.4
django-debug-toolbar==3.1.1
django-model-cache
django-extensions==3.0.9
django-filter==1.1.0
django-model-utils==4.0.0
django-notifications==0.1.dev0
django-notifications-hq==1.6.0
django-polymorphic==3.0.0
asgiref
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
coverage==5.3
datadog==0.39.0
decorator==4.4.2
gevent==20.9.0
greenlet
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
jsonfield==3.1.0
oauthlib==3.1.0
psycopg2-binary==2.8.6
public==2018.11.20
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-memcached==1.59
pytz==2020.4
redis==3.5.3
requests==2.25.0
requests-oauthlib==1.3.0
rjsmin==1.1.0
rsa==4.6
six==1.15.0
sqlparse==0.4.1
swapper==1.1.2.post1
uritemplate==3.0.1
urllib3==1.26.1
zope.event==4.5.0
zope.interface==5.2.0
confluent-kafka
JSON-log-formatter
